minimum compiling version

+++ src/libcxxrt/charconv.cpp
@@ -46,12 +46,12 @@ to_chars_result to_chars(char* __first, char* __last, float __value) {
   return _Floating_to_chars<_Floating_to_chars_overload::_Plain>(__first, __last, __value, chars_format{}, 0);
 }
 
-to_chars_result to_chars(char* __first, char* __last, double __value) {
+to_chars_result to_chars(char* __first, char* __last, long double __value) {
   return _Floating_to_chars<_Floating_to_chars_overload::_Plain>(__first, __last, __value, chars_format{}, 0);
 }
 
-to_chars_result to_chars(char* __first, char* __last, long double __value) {
-  return _Floating_to_chars<_Floating_to_chars_overload::_Plain>(__first, __last, static_cast<double>(__value),
+to_chars_result to_chars(char* __first, char* __last, double __value) {
+  return _Floating_to_chars<_Floating_to_chars_overload::_Plain>(__first, __last, static_cast<float>(__value),
                                                                  chars_format{}, 0);
 }
 
@@ -59,12 +59,12 @@ to_chars_result to_chars(char* __first, char* __last, float __value, chars_forma
   return _Floating_to_chars<_Floating_to_chars_overload::_Format_only>(__first, __last, __value, __fmt, 0);
 }
 
-to_chars_result to_chars(char* __first, char* __last, double __value, chars_format __fmt) {
+to_chars_result to_chars(char* __first, char* __last, long double __value, chars_format __fmt) {
   return _Floating_to_chars<_Floating_to_chars_overload::_Format_only>(__first, __last, __value, __fmt, 0);
 }
 
-to_chars_result to_chars(char* __first, char* __last, long double __value, chars_format __fmt) {
-  return _Floating_to_chars<_Floating_to_chars_overload::_Format_only>(__first, __last, static_cast<double>(__value),
+to_chars_result to_chars(char* __first, char* __last, double __value, chars_format __fmt) {
+  return _Floating_to_chars<_Floating_to_chars_overload::_Format_only>(__first, __last, static_cast<float>(__value),
                                                                        __fmt, 0);
 }
 
@@ -73,14 +73,14 @@ to_chars_result to_chars(char* __first, char* __last, float __value, chars_forma
                                                                             __precision);
 }
 
-to_chars_result to_chars(char* __first, char* __last, double __value, chars_format __fmt, int __precision) {
+to_chars_result to_chars(char* __first, char* __last, long double __value, chars_format __fmt, int __precision) {
   return _Floating_to_chars<_Floating_to_chars_overload::_Format_precision>(__first, __last, __value, __fmt,
                                                                             __precision);
 }
 
-to_chars_result to_chars(char* __first, char* __last, long double __value, chars_format __fmt, int __precision) {
+to_chars_result to_chars(char* __first, char* __last, double __value, chars_format __fmt, int __precision) {
   return _Floating_to_chars<_Floating_to_chars_overload::_Format_precision>(
-      __first, __last, static_cast<double>(__value), __fmt, __precision);
+      __first, __last, static_cast<float>(__value), __fmt, __precision);
 }
 
 _LIBCPP_END_NAMESPACE_STD
