+++ src/libcxxrt/include/ryu/d2s.h
@@ -53,7 +53,7 @@ inline constexpr int __DOUBLE_BIAS = 1023;
 inline constexpr int __DOUBLE_POW5_INV_BITCOUNT = 122;
 inline constexpr int __DOUBLE_POW5_BITCOUNT = 121;
 
-[[nodiscard]] to_chars_result __d2s_buffered_n(char* const _First, char* const _Last, const double __f, const chars_format _Fmt);
+[[nodiscard]] to_chars_result __d2s_buffered_n(char* const _First, char* const _Last, const long double __f, const chars_format _Fmt);
 
 _LIBCPP_END_NAMESPACE_STD
